feat: rename to cloud-ip-ranges.ts
