+++ src/lib/server.ts
@@ -3,7 +3,7 @@ import { initRedisClient } from './redis/client.js';
 import { probeOverride, probeIpLimit, initWsServer } from './ws/server.js';
 import { getMetricsAgent } from './metrics.js';
 import { populateMemList as populateMemMalwareList } from './malware/client.js';
-import { populateMemList as populateMemIpRangesList } from './ip-ranges.js';
+import { populateMemList as populateMemIpRangesList } from './cloud-ip-ranges.js';
 import { populateMemList as populateIpWhiteList } from './geoip/whitelist.js';
 import { populateCitiesList } from './geoip/city-approximation.js';
 import { reconnectProbes } from './ws/helper/reconnect-probes.js';
